rgw: dissect basic AWSv4's credentials parsing into separated function.
