+++ src/rgw/rgw_common.h
@@ -1694,7 +1694,6 @@ inline ostream& operator<<(ostream& out, const rgw_obj_index_key &o) {
 
 struct rgw_aws4_auth {
   string date;
-  string expires;
   string credential;
   string signedheaders;
   string signed_hdrs;
